add product

+++ src/main/java/com/tola/dev/phoneShope/service/BrandService.java
@@ -9,8 +9,8 @@ import com.tola.dev.phoneShope.entity.Brands;
 
 public interface BrandService {
 	Brands CreateBrand(Brands brands);
-	Brands getById(Integer id);
-	Brands updateById(Integer id , Brands brandsUpdateBrands);
+	Brands getById(Long id);
+	Brands updateById(Long id , Brands brandsUpdateBrands);
 	List<Brands> getBrands();
 	List<Brands> getBrands(String name);
 	List<Brands> getBrands(Map<String, String> params);
